+++ src/templates/Javadoc/Entities.module.scss
@@ -1,0 +1,48 @@
+@import 'variables';
+
+.property {
+    padding: $spacer / 2;
+    border: 1px solid $brand-grey-lighter;
+    margin-bottom: $spacer;
+    border-radius: $border-radius;
+
+    > p {
+        margin-bottom: $spacer;
+    }
+}
+
+.subHeading {
+    font-size: $font-size-base;
+    border-bottom: 1px solid $brand-grey-lighter;
+    padding-bottom: $spacer / 4;
+    margin-bottom: $spacer / $line-height;
+    color: $brand-grey;
+}
+
+.propertyName {
+    font-size: $font-size-large;
+    font-family: $font-family-monospace;
+    margin-bottom: $spacer / 2;
+    margin-top: 0;
+}
+
+.parameter {
+    margin-top: $spacer / 4;
+    margin-bottom: $spacer / 2;
+}
+
+.parameterType {
+    font-size: $font-size-small;
+    color: $brand-grey;
+    background: rgba($brand-blue, .2);
+    display: inline-block;
+    padding: 0 $spacer / 5;
+    border-radius: $border-radius;
+    vertical-align: middle;
+}
+
+.param {
+    display: inline-block;
+    margin-right: $spacer / 4;
+    font-weight: 600;
+}
