+++ src/templates/Javadoc/index.module.scss
@@ -1,31 +1,5 @@
 @import 'variables';
 
-.property {
-    padding: $spacer / 2;
-    border: 1px solid $brand-grey-lighter;
-    margin-bottom: $spacer;
-    border-radius: $border-radius;
-
-    > p {
-        margin-bottom: $spacer;
-    }
-}
-
-.subHeading {
-    font-size: $font-size-base;
-    border-bottom: 1px solid $brand-grey-lighter;
-    padding-bottom: $spacer / 4;
-    margin-bottom: $spacer / 4;
-    color: $brand-grey;
-}
-
-.propertyName {
-    font-size: $font-size-large;
-    font-family: $font-family-monospace;
-    margin-bottom: $spacer / 2;
-    margin-top: 0;
-}
-
 .namespace {
     font-size: $font-size-base;
     font-family: $font-family-monospace;
